fix(auth): fix import path for CreateUserDTO after moving to users module
